requirements.txt
python-telegram-bot>=20.0,<21
python-dotenv>=1.0.0
SQLAlchemy>=2.0,<3
aiosqlite>=0.19.0
asyncpg>=0.29.0
